Edits to file names/descriptions.
I tried to improve the interpretability of output names (files, rows/columns in them) where I thought it could be done.

Obviously, some of this is personal style, though I tried to keep that out of these comments.

In general, try to be as precise as you can (where you have the space). Recall that you cannot answer to Simon when he reads it. There is always a real possibility that he may think you produced a different statistic than you actually did.