Please rewrite this section
Please extract the dates directly from master.sliced as I show in the commented example. You may want to do it in a loop over the ID to make sure the order of the entries in the master.sliced df does not affect your result.

Please don't hand code it like this. There is substantial room for trouble e.g. orders can change, dates can be changed on the site (but remain untouched in your results) etc.